test(integration-tests): cover navigation flow against main
